+++ Sources/KankodaKit/Extensions/View+Extensions.swift
@@ -1,0 +1,19 @@
+//
+//  View+Extensions.swift
+//  Kankodakit
+//
+//  Created by Daniel Saidi on 2023-08-25.
+//  Copyright © 2023 Daniel Saidi. All rights reserved.
+//
+
+import SwiftUI
+
+extension View {
+    
+    public func fullWidthContent(
+        alignment: Alignment = .leading
+    ) -> some View {
+        self.frame(maxWidth: .infinity, alignment: .leading)
+            .contentShape(Rectangle())
+    }
+}